requirements.txt
pypdf
langchain
langchain-openai
langchain-community
langchain-core
python-dotenv